Updated Password Checker game based on Kyle's feedback
