+++ src/features/passwordChecker/components/PasswordInput.tsx
@@ -3,20 +3,18 @@ import TextField from '@mui/material/TextField';
 
 interface PasswordInputProps {
   password: string;
-  setPassword: (password: string) => void;
+  setPassword: (newPassword: string) => void;
 }
 
 const PasswordInput: React.FC<PasswordInputProps> = ({ password, setPassword }) => {
   return (
     <TextField
+      fullWidth
+      label="Enter Password"
       variant="outlined"
       type="password"
-      fullWidth
       value={password}
       onChange={(e) => setPassword(e.target.value)}
-      placeholder="Enter password"
-      inputProps={{ style: { color: '#4a4a4a' } }}  // Darker gray text
-      margin="normal"
     />
   );
 };
